refactor: edit squad settings integration
