+++ packages/shared/src/components/squads/settings/common.tsx
@@ -1,12 +1,7 @@
 import React, { ReactNode } from 'react';
-import { subDays } from 'date-fns';
 import classed from '../../../lib/classed';
 import { BlockIcon, LockIcon, TimerIcon } from '../../icons';
 import { IconSize } from '../../Icon';
-import {
-  PublicSquadRequest,
-  PublicSquadRequestStatus,
-} from '../../../graphql/sources';
 
 export const StatusDescription = classed(
   'p',
@@ -25,26 +20,4 @@ export const badge: Record<SquadStatus, ReactNode> = {
   [SquadStatus.Pending]: <TimerIcon size={IconSize.Small} />,
   [SquadStatus.Rejected]: <BlockIcon size={IconSize.Small} />,
   [SquadStatus.Approved]: null,
-};
-
-const remoteStatusMap: Record<PublicSquadRequestStatus, SquadStatus> = {
-  [PublicSquadRequestStatus.Approved]: SquadStatus.Approved,
-  [PublicSquadRequestStatus.Rejected]: SquadStatus.Rejected,
-  [PublicSquadRequestStatus.Pending]: SquadStatus.Pending,
-};
-
-export const getSquadStatus = (request: PublicSquadRequest): SquadStatus => {
-  if (!request) {
-    return SquadStatus.InProgress;
-  }
-
-  if (request.status === PublicSquadRequestStatus.Rejected) {
-    const fourteenDaysAgo = subDays(new Date(), 14);
-    const requestDate = new Date(request.createdAt);
-    return fourteenDaysAgo < requestDate
-      ? SquadStatus.Rejected
-      : SquadStatus.InProgress;
-  }
-
-  return remoteStatusMap[request.status];
 };
